Properly build include/exclude regex based on the path/notPath and name/notName parts.
Previously, parts would be simply added to the include/exclude regex array of the Finder. This means that having a ".txt" name and a "path" path would filter in any file in "path" OR any file with the ".txt" extension instead of any file in "path" AND with the ".txt" extension.